add new practices

+++ 06-Connecting-to-a-database-with-Spring-Data-JPA(H2)/demo/src/main/java/com/example/demo/repository/ProductRepository.java
@@ -1,8 +1,8 @@
 package com.example.demo.repository;
 
-import com.example.demo.model.Product;
+import com.example.demo.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
-
+public interface ProductRepository extends JpaRepository<Product, Long> {
+    // Bạn có thể thêm custom method nếu muốn
 }
